+++ lib/components/FieldLabel/FieldLabel.docs.tsx
@@ -10,18 +10,23 @@ import {
   Box,
   Hidden,
   Alert,
-  List,
+  Notice,
 } from '../';
 import { Placeholder } from '../../playroom/components';
+import { ReactNode } from 'markdown-to-jsx/node_modules/@types/react';
+
+const Container = ({ children }: { children: ReactNode }) => (
+  <Box maxWidth="xsmall">{children}</Box>
+);
 
 const docs: ComponentDocs = {
   category: 'Content',
   Example: () =>
     source(
-      <Box style={{ maxWidth: 400 }}>
-        <Stack space="small">
+      <Box maxWidth="xsmall">
+        <Stack space="xsmall">
           <FieldLabel
-            htmlFor="field"
+            htmlFor="example"
             label="Label"
             secondaryLabel={
               <>
@@ -34,7 +39,7 @@ const docs: ComponentDocs = {
               </TextLink>
             }
           />
-          <Box id="field">
+          <Box id="example">
             <Placeholder height={40} />
           </Box>
         </Stack>
@@ -55,18 +60,151 @@ const docs: ComponentDocs = {
       associate the label with its associated field.
     </Text>
   ),
-  alternatives: [],
+  alternatives: [
+    {
+      name: 'FieldMessage',
+      description: 'For displaying messages below a custom field.',
+    },
+  ],
   additional: [
     {
-      label: 'See also',
+      label: 'Label',
+      Container,
+      description: <Text>Sets the primary label of field.</Text>,
+      Example: () =>
+        source(
+          <Stack space="xsmall">
+            <FieldLabel htmlFor="labelExample" label="Label" />
+            <Box id="labelExample">
+              <Placeholder height={40} />
+            </Box>
+          </Stack>,
+        ),
+    },
+    {
+      label: 'Secondary label',
+      Container,
       description: (
-        <List space="large">
-          <Text tone="secondary">
-            <TextLink href="/components/FieldMessage">FieldMessage</TextLink>{' '}
-            &mdash; For displaying messages below a custom field.
+        <>
+          <Text>
+            Provide additional context, typically used to indicate optionality
+            of a field.
           </Text>
-        </List>
+          <Notice tone="info">
+            <Text>
+              Only displayed when provided in addition to a{' '}
+              <TextLink href="#label">label</TextLink>.
+            </Text>
+          </Notice>
+        </>
       ),
+      Example: () =>
+        source(
+          <Stack space="xsmall">
+            <FieldLabel
+              htmlFor="secondaryExample"
+              label="Label"
+              secondaryLabel="Secondary label"
+            />
+            <Box id="secondaryExample">
+              <Placeholder height={40} />
+            </Box>
+          </Stack>,
+        ),
+    },
+    {
+      label: 'Tertiary label',
+      Container,
+      description: (
+        <>
+          <Text>
+            Provide further context, typically used for providing assistance
+            with a field.
+          </Text>
+          <Notice tone="info">
+            <Text>
+              Only displayed when provided in addition to a{' '}
+              <TextLink href="#label">label</TextLink>.
+            </Text>
+          </Notice>
+        </>
+      ),
+      Example: () =>
+        source(
+          <Stack space="xsmall">
+            <FieldLabel
+              htmlFor="tertiaryExample"
+              label="Label"
+              tertiaryLabel={
+                <TextLink href="#">
+                  Tertiary<Hidden below="tablet"> label</Hidden>
+                </TextLink>
+              }
+            />
+            <Box id="tertiaryExample">
+              <Placeholder height={40} />
+            </Box>
+          </Stack>,
+        ),
+    },
+    {
+      label: 'Additional description',
+      Container,
+      description: (
+        <>
+          <Text>
+            Additional context can be provided with a{' '}
+            <Strong>description</Strong>. This will display below the field
+            label.
+          </Text>
+          <Alert tone="info">
+            <Text>
+              It is strongly recommended that a <Strong>descriptionId</Strong>{' '}
+              is also passed in, allowing the description to be associated with
+              your custom field via <Strong>aria-describedby</Strong> prop. This
+              will allow a screen reader to provide this additional context upon
+              focusing the field.
+            </Text>
+          </Alert>
+        </>
+      ),
+      Example: () =>
+        source(
+          <Stack space="xsmall">
+            <FieldLabel
+              htmlFor="descriptionExample"
+              label="Label"
+              description="Extra information about the field"
+              descriptionId="fieldDescription"
+            />
+            <Box id="descriptionExample" aria-describedby="fieldDescription">
+              <Placeholder height={40} />
+            </Box>
+          </Stack>,
+        ),
+    },
+    {
+      label: 'Disabled field',
+      Container,
+      description: (
+        <Text>
+          Setting the <Strong>disabled</Strong> prop to <Strong>true</Strong>{' '}
+          dims the field label.
+        </Text>
+      ),
+      Example: () =>
+        source(
+          <Stack space="xsmall">
+            <FieldLabel
+              htmlFor="disabledExample"
+              label="Label"
+              disabled={true}
+            />
+            <Box id="disabledExample">
+              <Placeholder height={40} />
+            </Box>
+          </Stack>,
+        ),
     },
   ],
 };