+++ lib/components/Radio/Radio.screenshots.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { ComponentScreenshot } from '../../../site/src/types';
-import { Badge, Radio, Text } from '../';
+import { Badge, Radio, Text, Stack } from '../';
 
 export const screenshots: ComponentScreenshot = {
   screenshotWidths: [320],
@@ -28,13 +28,22 @@ export const screenshots: ComponentScreenshot = {
     {
       label: 'Disabled Radio Button',
       Example: ({ id, handler }) => (
-        <Radio
-          id={id}
-          disabled={true}
-          checked={false}
-          onChange={handler}
-          label="Label"
-        />
+        <Stack space="gutter">
+          <Radio
+            id={`${id}_1`}
+            disabled={true}
+            checked={false}
+            onChange={handler}
+            label="Disabled unchecked"
+          />
+          <Radio
+            id={`${id}_2`}
+            disabled={true}
+            checked={true}
+            onChange={handler}
+            label="Disabled checked"
+          />
+        </Stack>
       ),
     },
     {